Add missing demo app changes for extension detection UI
Brings the demo app changes from feat/fix-auto-confirm branch to the clean branch:

- Added ConfigurationForm.tsx component for provider selection and extension status
- Added SDKActions.tsx component with auto-confirm functionality
- Updated App.tsx to use ConfigurationForm before SDK instantiation
- Simplified Actions.tsx to use SDKActions component
- Added extension detection UI with useIsExtensionInstalled hook

This completes the missing 5 files from the original fix-auto-confirm branch
that provide the demo app UI for testing extension detection and auto-confirm.
