+++ examples/react-sdk-demo-app/src/App.tsx
@@ -1,6 +1,7 @@
 import { Routes, Route } from "react-router-dom";
 import { AddressType, PhantomProvider, type PhantomSDKConfig, type PhantomDebugConfig, DebugLevel, type DebugMessage, type ProviderType } from "@phantom/react-sdk";
 import { Actions } from "./Actions";
+import { ConfigurationForm } from "./ConfigurationForm";
 import { AuthCallback } from "./AuthCallback";
 import { useState, useCallback, useMemo } from "react";
 import { DebugContext, type DebugContextType } from "./contexts/DebugContext";
@@ -14,6 +15,9 @@ function App() {
   // Provider type state - this will control the SDK configuration
   const [providerType, setProviderType] = useState<ProviderType>("embedded");
   const [embeddedWalletType, setEmbeddedWalletType] = useState<"user-wallet" | "app-wallet">("user-wallet");
+  
+  // SDK instantiation state
+  const [sdkInstantiated, setSdkInstantiated] = useState(false);
 
   // Debug callback function for the provider
   const handleDebugMessage = useCallback((message: DebugMessage) => {
@@ -67,21 +71,47 @@ function App() {
     clearDebugMessages,
   }), [debugMessages, debugLevel, showDebug, setDebugLevel, setShowDebug, clearDebugMessages]);
 
+  // Auth callback always needs embedded config
+  const authConfig: PhantomSDKConfig = useMemo(() => ({
+    appName: "React SDK Demo App",
+    appLogo: "https://picsum.photos/200",
+    providerType: "embedded",
+    addressTypes: [AddressType.solana, AddressType.ethereum, AddressType.bitcoinSegwit, AddressType.sui],
+    solanaProvider: "web3js",
+    organizationId: import.meta.env.VITE_ORGANIZATION_ID || "your-organization-id",
+    apiBaseUrl: import.meta.env.VITE_API_BASE_URL || "https://api.phantom.app/v1/wallets",
+    embeddedWalletType: "user-wallet", // Auth callback is always for user wallet
+    authOptions: {
+      authUrl: import.meta.env.VITE_AUTH_URL || "https://connect.phantom.app",
+      redirectUrl: import.meta.env.VITE_REDIRECT_URL,
+    },
+    autoConnect: true,
+  }), []);
+
   return (
     <DebugContext.Provider value={debugContextValue}>
-      <PhantomProvider config={config} debugConfig={debugConfig}>
-        <Routes>
-          <Route path="/" element={
-            <Actions 
+      <Routes>
+        <Route path="/auth/callback" element={
+          <PhantomProvider config={authConfig} debugConfig={debugConfig}>
+            <AuthCallback />
+          </PhantomProvider>
+        } />
+        <Route path="/" element={
+          sdkInstantiated ? (
+            <PhantomProvider config={config} debugConfig={debugConfig}>
+              <Actions providerType={providerType} />
+            </PhantomProvider>
+          ) : (
+            <ConfigurationForm 
               providerType={providerType}
               setProviderType={setProviderType}
               embeddedWalletType={embeddedWalletType}
               setEmbeddedWalletType={setEmbeddedWalletType}
+              onCreateSDK={() => setSdkInstantiated(true)}
             />
-          } />
-          <Route path="/auth/callback" element={<AuthCallback />} />
-        </Routes>
-      </PhantomProvider>
+          )
+        } />
+      </Routes>
     </DebugContext.Provider>
   );
 }
